requirements.txt
dlib==19.24.6
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
